wip: recaptcha, package.json cleanup
