+++ apps/webapp/package.json
@@ -44,7 +44,7 @@
     "pino-pretty": "^10.3.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-google-recaptcha-v3": "^1.10.1",
+    "react-google-recaptcha": "^3.1.0",
     "react-hook-form": "^7.51.1",
     "react-hot-toast": "^2.4.1",
     "react-qr-code": "^2.0.12",
@@ -76,6 +76,9 @@
     "tailwind-merge": "^2.2.1",
     "tailwindcss": "^3.4.1",
     "tailwindcss-animate": "^1.0.7",
-    "typescript": "^5.3.3"
+    "typescript": "^5.3.3",
+    "@types/react-google-recaptcha": "^2.1.9",
+    "@types/pako": "^2.0.3",
+    "prettier-plugin-tailwindcss": "^0.5.12"
   }
 }
